+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Change Log
 
-## [1.8.5] - Build 292
+## [1.8.5] - Build 293
 
 - New
+  - [x] **Color Preview**
+    - You can now see color preview in css, scss, less, stylus and sass codes.
+    - No need to select the whole color.
+    - Enable/disable this feature from editor settings.
   - [x] **Smaller app size**
     - Reduced app size.
   - [x] **Updated icon pack**